Added SO=6
as a half-sized DX12_ME mode. Also possibly fixed bug with passing invalid vectors to MDegrain. Vertical size for DX12_ME looks like need to give integer number of blocks (mod 8 for 8x8 block size and may be mod 16 for 16x16 block size). The performance of SAD compute engine in MAnalyse is limiting factor now so it is planned to add DX12 compute shader SAD calculation in the next updates.