Added AreaMode for
MAnalyse. New params: AreaMode of 0 (default disabled) and 1 to 4 of diagonal additional searches of MVs and Mode-like final selection of most common. Also added SAD hinting to MDegrain with AMdiffSAD param (multiplier) of the mean abs difference of center block MV and area searched MVs. So for unstable blocks the SAD will be raised to have lower or none weight in MDegrain.